Transaction for p2p trading
This contract defines a Transaction struct to hold transaction information, a User struct to hold user information, and a EnergyTrading contract that enables energy trading between users. Users can buy and sell energy from each other, and transactions must be approved by the receiver before they are executed. The contract also includes a function to set the energy price and retrieve the list of allowed users and their transactions.